Expand diffing protocol for clarity
The current incarnation of the protocol says when the diffing is over,
but not when the elements from the diff are over. This works, but causes
the following woes:

- makes testing a whole lot more convoluted
- impossible to make a useful synchronous diff call (which could be
  useful for snapshotting, tests, etc.)
- makes it harder to enable some TCP optimizations based on protocol
  context

This patch fixes this by replacing the existing flow that basically
goes:

1. diff
2. stream updates and other stuff indistinctively in relay mode

into:

1. diff
2. stream diffed updates
3. back to relay mode

This allows the client to relate 'sync_done' as a message that has more
meaning and opens up a few more opportunities.