Allow to stream updates during diffing
Rather than the serving interrupting all updates and then de-duping them
by key with the diff results and *then* streaming it all at once, the
server relays the updates as they come in, and only stores the keys in
memory.

When the diff is done, it deduplicates following the usual mechanism and
only sends the differing entries that were *not* modified during the
diff back to the client. This logically gives the same result set, but
without delaying in-flight data.

The client has been extended accordingly.